+++ r-package/R/pareto_frontier.R
@@ -132,7 +132,7 @@ pareto_frontier <- function(r5r_core,
                             max_lts = 2,
                             draws_per_minute = 5L,
                             n_threads = Inf,
-                            verbose = TRUE,
+                            verbose = FALSE,
                             progress = TRUE) {