+++ r-package/R/travel_time_matrix.R
@@ -146,7 +146,7 @@ travel_time_matrix <- function(r5r_core,
                                max_lts = 2,
                                draws_per_minute = 5L,
                                n_threads = Inf,
-                               verbose = TRUE,
+                               verbose = FALSE,
                                progress = TRUE) {