+++ r-package/man/roxygen/templates/common_arguments.R
@@ -51,8 +51,9 @@
 #' information.
 #' @param n_threads An integer. The number of threads to use when running the
 #' router in parallel. Defaults to use all available threads (Inf).
-#' @param verbose A logical. Defaults to `TRUE`, showing detailed output
-#' messages. When `FALSE`, only `R5` error messages are shown.
+#' @param verbose A logical. Defaults to `FALSE`, showing only eventual ERROR
+#' messages. Setting `verbose=TRUE` shows detailed output, which can be useful
+#' for debugging.
 #' @param progress A logical. Whether to show a progress counter when running
 #' the router. Defaults to `TRUE`. Only works when `verbose` is set to `FALSE`,
 #' so the progress counter does not interfere with `R5`'s output messages.